Reading_and_categorizing_scanned_documents/requirements.txt
numpy >=1.16.0, <2.0.0
scipy >=1.4.0, <2.0.0
pillow >=9.2.0
matplotlib >=3.1.0
h5py >=3.1.0, <4.0.0
opencv >=4.5.0, <5.0.0
pypdfium2-team::pypdfium2_helpers >=4.0.0, <5.0.0
pyclipper >=1.2.0, <2.0.0
shapely >=1.6.0, <3.0.0
langdetect >=1.0.9, <2.0.0
rapidfuzz >=3.0.0, <4.0.0
huggingface_hub >=0.20.0, <1.0.0
matplotlib >=3.1.0
weasyprint >=55.0
defusedxml >=0.7.0
mplcursors >=0.3
unidecode >=1.0.0
tqdm >=4.30.0